# node-API/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ApolloX

host: localhost:3000
basePath: /

schemes:
  - http
  - https
consumes:
  - application/json
  - application/x-www-form-urlencoded
  - application/octet-stream
  - multipart/form-data
produces:
  - application/json

securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      The following syntax must be used in the 'Authorization' header :
      "Bearer {token}"

paths:
  /user:
    x-swagger-router-controller: main_controller
    get:
      description: User profile
      operationId: getUser
      security:
        - Bearer: []
      x-security-scopes:
        - user
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Message"
    put:
      description: Update User
      operationId: updateUser
      security:
        - Bearer: []
      x-security-scopes:
        - user
      parameters:
        - name: userUpdate
          in: body
          schema:
            properties:
              name:
                type: string
              email_id:
                type: string
              phone_no:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Message"

  /auth/login/otp:
    x-swagger-router-controller: main_controller
    post:
      description: Login User through OTP. Send ID token
      operationId: userLoginOtp
      parameters:
        - name: userLoginOtp
          in: body
          schema:
            properties:
              token:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Message"

  /auth/login/google:
    x-swagger-router-controller: main_controller
    post:
      description: Login User through google. Send ID token
      operationId: userLoginGoogle
      parameters:
        - name: userLoginGoogle
          in: body
          schema:
            properties:
              token:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Message"



# -----------------------------------------------------------
definitions:
  Message:
    required:
      - message
    properties:
      message:
        type: string

  User:
    properties:
      uid:
        type: string
      name:
        type: string
      email_id:
        type: string
      phone_no:
        type: string
      role:
        type: string
      create_time:
        type: string